pyproject.toml
[tool.poetry]
name = "auto-finance"
version = "0.1.0"
description = "portfolio management with genAI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
google-generativeai = "^0.8.3"
langgraph = "^0.2.60"
yfinance = "^0.2.36"
pandas = "^2.1.4"
numpy = "^1.26.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pydantic = "^2.5.2"
langchain = "^0.3.13"
langchain-openai = "^0.2.14"
transformers = "^4.47.1"
tensorflow = "2.16.2"
tf-keras = "2.16.0"
langchain-google-genai = "^2.0.7"
anthropic = "^0.42.0"
dependency-injector = "^4.44.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.myp]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.poetry.scripts]
auto_finance = "auto_finance.cli"